test: user login ok
