+++ pubspec.yaml
@@ -42,6 +42,8 @@ dependencies:
   firebase_core: ^3.12.1
   cloud_firestore: ^5.6.5
   firebase_storage: ^12.4.4
+  email_validator: ^3.0.0
+  firebase_auth: ^5.5.0
